Rename ErrorAlert and Loader to be kebab-case, add missing 'use client' declarations, update imports
