+++ client/src/components/ui/error-alert.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { ExclamationTriangleIcon } from '@heroicons/react/24/outline';
 import {
   Button,
